Add prefetching at the packet struct
Leaves open which level of temporal consistency to use for the access
and exposes an unconditionally avaiable enumeration to differentiate
them (potentially useful should prefetching be available for other
architectures).

This also performs prefetching at level Time1 (per documentation into L2
but not L1) for each received packet when the method is enabled. That
was determined fastest in the test cluster.