Unsafe block cleanup (#17)
* memory: get_phys_addr() is not unsafe

There's nothing in get_phys_addr() that requires the method to be
annotated as unsafe.

* memory: Make get_virt_addr() function signature safe

The get_virt_addr() function performs an unsafe operation, but by
asserting on the buffer id, the method can be marked as safe.

* ixgbe: Eliminate unsafe block from rx_batch()

There's no need to keep the whole Packet construction unsafe now that
both get_phys_addr() and get_virt_addr() are safe. Wrap the call to
read_volatile() with unsafe block, and make Packet construction safe.