Fix IPv4 checksum calculation and wait until all packets are queued
The previous checksum calculation was incorrect which caused packets
to be filtered out by Virtio bridges.
Also wait until all packets are queued to match the upstream pktgen
example application.
Finally the source MAC is changed so it isn't a multicast address
anymore as that is also filtered out by Virtio bridges.